+++ tests/util.go
@@ -4,6 +4,6 @@ import (
 	"math"
 )
 
-func IsEquals(a, b float64) bool {
-	return math.Abs(a-b) < 0.00001
+func IsEquals(a, b float32) bool {
+	return math.Abs(float64(a-b)) < 0.00001
 }
